+++ hooks/src/lib/hooks/index.ts
@@ -1,6 +1,6 @@
-export { default as useCardIssuer } from './InputField/useInputCardIssuer';
-export { default as useCardNumber } from './InputField/useInputCardNumber';
-export { default as useCVCNumber } from './InputField/useInputCVCNumber';
-export { default as useExpiryDate } from './InputField/useInputExpiryDate';
-export { default as useOwnerName } from './InputField/useInputOwnerName';
-export { default as usePasswordPrefix } from './InputField/useInputPasswordPrefix';
+export { default as useCardIssuer } from './useInputCardIssuer';
+export { default as useCardNumber } from './useInputCardNumber';
+export { default as useCVCNumber } from './useInputCVCNumber';
+export { default as useExpiryDate } from './useInputExpiryDate';
+export { default as useOwnerName } from './useInputOwnerName';
+export { default as usePasswordPrefix } from './useInputPasswordPrefix';
