+++ hooks/src/lib/hooks/useInputPasswordPrefix.ts
@@ -1,29 +1,29 @@
 import { useState } from 'react';
 import { getInputStatus, useInput } from './useInput';
-import { ERROR_MESSAGE } from '../../shared/errorMessages';
-import validator from '../../shared/utils/validator/validator';
-import { Status } from '../../shared/types';
-import { VALID_LENGTH } from '../../shared/options';
+import { ERROR_MESSAGE } from '../shared/errorMessages';
+import validator from '../shared/utils/validator/validator';
+import { Status } from '../shared/types';
+import { VALID_LENGTH } from '../shared/options';
 
-type UseInputCardNumberReturn = [
+type UseInputPasswordPrefixReturn = [
   value: string,
   status: Status,
   errorMessage: string,
-  handleChange: (value: string, validLength?: number) => void,
+  handleChange: (value: string) => void,
   handleBlur: () => void
 ];
 
-const useInputCardNumber = (): UseInputCardNumberReturn => {
+const useInputPasswordPrefix = (): UseInputPasswordPrefixReturn => {
   const { value, status, setValue, setStatus } = useInput('');
   const [errorMessage, setErrorMessage] = useState<string>('');
 
-  const handleChange = (value: string, validLength: number = VALID_LENGTH.cardNumber) => {
-    //  status 업데이트
-    setStatus(getInputStatus(value, validLength));
+  const handleChange = (value: string) => {
+    // status 업데이트
+    const status = getInputStatus(value, VALID_LENGTH.passwordPrefix);
 
-    // Default가 아닌 경우 : Error 검사
+    // Default인 경우 : Error 검사
     if (status !== 'default') {
-      const [isValid, errorMessage] = validator.cardNumber.isValidInput(value);
+      const [isValid, errorMessage] = validator.passwordPrefix.isValidInput(value);
 
       // Error인 경우 : 에러 발생
       if (!isValid) {
@@ -35,6 +35,7 @@ const useInputCardNumber = (): UseInputCardNumberReturn => {
 
     // Error가 아닌 경우 : 값 업데이트
     setValue(value);
+    setStatus(status);
     setErrorMessage('');
   };
 
@@ -42,11 +43,11 @@ const useInputCardNumber = (): UseInputCardNumberReturn => {
     // 미완성인 경우 : Error 상태로 판단
     if (['default', 'pending'].includes(status)) {
       setStatus('error');
-      setErrorMessage(ERROR_MESSAGE.cardNumber.isNotFulfilled);
+      setErrorMessage(ERROR_MESSAGE.passwordPrefix.isNotFulfilled);
     }
   };
 
   return [value, status, errorMessage, handleChange, handleBlur];
 };
 
-export default useInputCardNumber;
+export default useInputPasswordPrefix;
